moved deleteUser authentication to service

+++ domain/services/userService.js
@@ -49,37 +49,43 @@ var userService = {
       }
     })
   },
-  deleteUser: (phone, callback) => {
+  deleteUser: (phone, tokenId, callback) => {
     userRepository.read(phone, (error, userData) => {
       if (!error && userData) {
-        userRepository.delete(userData.phone, (error) => {
-          if (!error) {
-            var userCheckIds = typeof(userData.checks) == 'object' && userData.checks instanceof Array ? userData.checks : []
-            var nChecksToDelete = userCheckIds.length
-            if (0 < nChecksToDelete) {
-              var nChecksDeleted = 0
-              var deletionErrors = []
-              userCheckIds.forEach((checkId) => {
-                checkRepository.delete(checkId, (error) => {
-                  if (error) {
-                    deletionErrors.push(error)
-                  }
-                  nChecksDeleted++;
-                  if (nChecksDeleted == nChecksToDelete) {
-                    if (!deletionErrors.length) {
-                      callback(false)
-                    } else {
-                      callback({ deletionErrors })
-                    }
-                  }
-    
-                })
-              })
-            } else {
-              callback(false)
-            }
+        tokenVerifier.matchTokenAndPhone(tokenId, userData.phone, (isMatching) => {
+          if (isMatching) {
+            userRepository.delete(userData.phone, (error) => {
+              if (!error) {
+                var userCheckIds = typeof(userData.checks) == 'object' && userData.checks instanceof Array ? userData.checks : []
+                var nChecksToDelete = userCheckIds.length
+                if (0 < nChecksToDelete) {
+                  var nChecksDeleted = 0
+                  var deletionErrors = []
+                  userCheckIds.forEach((checkId) => {
+                    checkRepository.delete(checkId, (error) => {
+                      if (error) {
+                        deletionErrors.push(error)
+                      }
+                      nChecksDeleted++;
+                      if (nChecksDeleted == nChecksToDelete) {
+                        if (!deletionErrors.length) {
+                          callback(false)
+                        } else {
+                          callback({ deletionErrors })
+                        }
+                      }
+        
+                    })
+                  })
+                } else {
+                  callback(false)
+                }
+              } else {
+                callback(error)
+              }
+            })
           } else {
-            callback(error)
+            callback({ error: 'invalid token, unauthorized' }, null)
           }
         })
       } else {
